+++ src/main/java/io/github/chasehuegel/skilling/engine/ui/SkillMenuBuilder.java
@@ -170,7 +170,7 @@ public final class SkillMenuBuilder {
     private int getLevelForXp(SkillDefinition skill, long xp) {
         for (int level = 1; level <= skill.maxLevel(); level++) {
             double required = skill.progression().evaluator().evaluate(level, 0);
-            if (xp < required) return level - 1;
+            if (xp < (long) required) return level - 1;
         }
         return skill.maxLevel();
     }
